Crate rustc_mir_build

Construction of MIR from HIR.

Modules

build 🔒

check_tail_calls 🔒

check_unsafety 🔒

errors 🔒

lints

thir 🔒

The MIR is built from some typed high-level IR (THIR). This section defines the THIR along with a trait for accessing it. The intention is to allow MIR construction to be unit-tested and separated from the Rust source and compiler data structures.

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Functions

provide