rustc_mir_build::build::expr::as_place
Function strip_prefix

fn strip_prefix<'a, 'tcx>(
    base_ty: Ty<'tcx>,
    projections: &'a [PlaceElem<'tcx>],
    prefix_projections: &[Projection<'tcx>],
) -> impl Iterator<Item = PlaceElem<'tcx>> + 'a

Returns projections remaining after stripping an initial prefix of HIR projections.

Supports only HIR projection kinds that represent a path that might be captured by a closure or a coroutine, i.e., an Index or a Subslice projection kinds are unsupported.