Struct rustc_mir_build::check_unsafety::UnsafetyVisitor

struct UnsafetyVisitor<'a, 'tcx> {
    tcx: TyCtxt<'tcx>,
    thir: &'a Thir<'tcx>,
    hir_context: HirId,
    safety_context: SafetyContext,
    body_unsafety: BodyUnsafety,
    body_target_features: &'tcx [Symbol],
    assignment_info: Option<(Ty<'tcx>, Span)>,
    in_union_destructure: bool,
    param_env: ParamEnv<'tcx>,
    inside_adt: bool,
}

Fields
tcx: TyCtxt<'tcx>thir: &'a Thir<'tcx>hir_context: HirId

The HirId of the current scope, which would be the HirId of the current HIR node, modulo adjustments. Used for lint levels.

safety_context: SafetyContext

The current “safety context”. This notably tracks whether we are in an unsafe block, and whether it has been used.

body_unsafety: BodyUnsafetybody_target_features: &'tcx [Symbol]

The #[target_feature] attributes of the body. Used for checking calls to functions with #[target_feature] (RFC 2396).

assignment_info: Option<(Ty<'tcx>, Span)>

When inside the LHS of an assignment to a field, this is the type of the LHS and the span of the assignment expression.

in_union_destructure: boolparam_env: ParamEnv<'tcx>inside_adt: bool
Implementations

impl<'tcx> UnsafetyVisitor<'_, 'tcx>

fn in_safety_context(
&mut self,
safety_context: SafetyContext,
f: impl FnOnce(&mut Self)
)

fn requires_unsafe(&mut self, span: Span, kind: UnsafeOpKind)

fn warn_unused_unsafe(
&self,
hir_id: HirId,
block_span: Span,
enclosing_unsafe: Option<(Span, &'static str)>
)

fn unsafe_op_in_unsafe_fn_allowed(&self) -> bool

Whether the unsafe_op_in_unsafe_fn lint is allowed at the current HIR node.