Module rustc_mir_build::thir::cx

This module contains the functionality to convert from the wacky tcx data structures into the THIR. The builder is generally ignorant of the tcx, etc., and instead goes through the Cx for most of its work.

Modules

block 🔒

expr 🔒

Structs

Cx 🔒

Functions

thir_body 🔒

thir_tree 🔒