Struct rustc_mir_build::thir::cx::Cx

struct Cx<'tcx> {
    tcx: TyCtxt<'tcx>,
    thir: Thir<'tcx>,
    param_env: ParamEnv<'tcx>,
    region_scope_tree: &'tcx ScopeTree,
    typeck_results: &'tcx TypeckResults<'tcx>,
    rvalue_scopes: &'tcx RvalueScopes,
    adjustment_span: Option<(HirId, Span)>,
    body_owner: DefId,
}

Fields
tcx: TyCtxt<'tcx>thir: Thir<'tcx>param_env: ParamEnv<'tcx>region_scope_tree: &'tcx ScopeTreetypeck_results: &'tcx TypeckResults<'tcx>rvalue_scopes: &'tcx RvalueScopesadjustment_span: Option<(HirId, Span)>

When applying adjustments to the expression with the given HirId, use the given Span, instead of the usual span. This is used to assign the span of an overall method call (e.g. my_val.foo()) to the adjustment expressions for the receiver.

body_owner: DefId

The DefId of the owner of this body.

Implementations

impl<'tcx> Cx<'tcx>

pub(crate) fn mirror_block(&mut self, block: &'tcx Block<'tcx>) -> BlockId

fn mirror_stmts(
&mut self,
block_id: ItemLocalId,
stmts: &'tcx [Stmt<'tcx>]
) -> Box<[StmtId]>

impl<'tcx> Cx<'tcx>

pub(crate) fn mirror_expr(&mut self, expr: &'tcx Expr<'tcx>) -> ExprId

pub(crate) fn mirror_exprs(&mut self, exprs: &'tcx [Expr<'tcx>]) -> Box<[ExprId]>

pub(super) fn mirror_expr_inner(&mut self, hir_expr: &'tcx Expr<'tcx>) -> ExprId

fn apply_adjustment(
&mut self,
hir_expr: &'tcx Expr<'tcx>,
expr: Expr<'tcx>,
adjustment: &Adjustment<'tcx>,
span: Span
) -> Expr<'tcx>

fn mirror_expr_cast(
&mut self,
source: &'tcx Expr<'tcx>,
temp_lifetime: Option<Scope>,
span: Span
) -> ExprKind<'tcx>

Lowers a cast expression.

Dealing with user type annotations is left to the caller.

fn make_mirror_unadjusted(&mut self, expr: &'tcx Expr<'tcx>) -> Expr<'tcx>

fn user_substs_applied_to_res(
&mut self,
hir_id: HirId,
res: Res
) -> Option<Box<CanonicalUserType<'tcx>>>

fn method_callee(
&mut self,
expr: &Expr<'_>,
span: Span,
overloaded_callee: Option<(DefId, SubstsRef<'tcx>)>
) -> Expr<'tcx>

fn convert_arm(&mut self, arm: &'tcx Arm<'tcx>) -> ArmId

fn convert_path_expr(
&mut self,
expr: &'tcx Expr<'tcx>,
res: Res
) -> ExprKind<'tcx>

fn convert_var(&mut self, var_hir_id: HirId) -> ExprKind<'tcx>

fn overloaded_operator(
&mut self,
expr: &'tcx Expr<'tcx>,
args: Box<[ExprId]>
) -> ExprKind<'tcx>

fn overloaded_place(
&mut self,
expr: &'tcx Expr<'tcx>,
place_ty: Ty<'tcx>,
overloaded_callee: Option<(DefId, SubstsRef<'tcx>)>,
args: Box<[ExprId]>,
span: Span
) -> ExprKind<'tcx>

fn convert_captured_hir_place(
&mut self,
closure_expr: &'tcx Expr<'tcx>,
place: HirPlace<'tcx>
) -> Expr<'tcx>

fn capture_upvar(
&mut self,
closure_expr: &'tcx Expr<'tcx>,
captured_place: &'tcx CapturedPlace<'tcx>,
upvar_ty: Ty<'tcx>
) -> Expr<'tcx>

fn field_refs(&mut self, fields: &'tcx [ExprField<'tcx>]) -> Box<[FieldExpr]>

Converts a list of named fields (i.e., for struct-like struct/enum ADTs) into FieldExpr.

impl<'tcx> Cx<'tcx>

fn new(tcx: TyCtxt<'tcx>, def: WithOptConstParam<LocalDefId>) -> Cx<'tcx>

fn pattern_from_hir(&mut self, p: &Pat<'_>) -> Box<Pat<'tcx>>

fn closure_env_param(
&self,
owner_def: LocalDefId,
owner_id: HirId
) -> Option<Param<'tcx>>

fn explicit_params<'a>(
&'a mut self,
owner_id: HirId,
fn_decl: &'tcx FnDecl<'tcx>,
body: &'tcx Body<'tcx>
) -> impl Iterator<Item = Param<'tcx>> + 'a