Module rustc_mir_build::thir::pattern

Validation of patterns/matches.

Modules

check_match ğŸ”’

const_to_pat ğŸ”’

deconstruct_pat ğŸ”’

super::usefulness explains most of what is happening in this file. As explained there, values and patterns are made from constructors applied to fields. This file defines a Constructor enum, a Fields struct, and various operations to manipulate them and convert them from/to patterns.

usefulness ğŸ”’

Note: tests specific to this file can be found in:

Macros

ClonePatternFoldableImpls ğŸ”’

Structs

PatCtxt ğŸ”’

Enums

PatternError ğŸ”’

Traits

PatternFoldable ğŸ”’

PatternFolder ğŸ”’

Functions

compare_const_vals ğŸ”’

pat_from_hir ğŸ”’