rustc_mir_build::thir::pattern::check_match
Struct MatchVisitor

struct MatchVisitor<'p, 'tcx> {
    tcx: TyCtxt<'tcx>,
    param_env: ParamEnv<'tcx>,
    typeck_results: &'tcx TypeckResults<'tcx>,
    thir: &'p Thir<'tcx>,
    lint_level: HirId,
    let_source: LetSource,
    pattern_arena: &'p TypedArena<DeconstructedPat<'p, 'tcx>>,
    dropless_arena: &'p DroplessArena,
    error: Result<(), ErrorGuaranteed>,
}

Fields
tcx: TyCtxt<'tcx>param_env: ParamEnv<'tcx>typeck_results: &'tcx TypeckResults<'tcx>thir: &'p Thir<'tcx>lint_level: HirIdlet_source: LetSourcepattern_arena: &'p TypedArena<DeconstructedPat<'p, 'tcx>>dropless_arena: &'p DroplessArenaerror: Result<(), ErrorGuaranteed>

Tracks if we encountered an error while checking this body. That the first function to report it stores it here. Some functions return Result to allow callers to short-circuit on error, but callers don’t need to store it here again.

Implementations

impl<'p, 'tcx> MatchVisitor<'p, 'tcx>

fn with_let_source(&mut self, let_source: LetSource, f: impl FnOnce(&mut Self))

fn with_lint_level<T>( &mut self, new_lint_level: LintLevel, f: impl FnOnce(&mut Self) -> T, ) -> T

fn visit_land( &mut self, ex: &'p Expr<'tcx>, accumulator: &mut Vec<Option<(Span, RefutableFlag)>>, ) -> Result<(), ErrorGuaranteed>

Visit a nested chain of &&. Used for if-let chains. This must call visit_expr on the subexpressions we are not handling ourselves.

fn visit_land_rhs( &mut self, ex: &'p Expr<'tcx>, ) -> Result<Option<(Span, RefutableFlag)>, ErrorGuaranteed>

Visit the right-hand-side of a &&. Used for if-let chains. Returns Some if the expression was ultimately a let ... = ..., and None if it was a normal boolean expression. This must call visit_expr on the subexpressions we are not handling ourselves.

fn lower_pattern( &mut self, cx: &PatCtxt<'p, 'tcx>, pat: &'p Pat<'tcx>, ) -> Result<&'p DeconstructedPat<'p, 'tcx>, ErrorGuaranteed>

fn is_known_valid_scrutinee(&self, scrutinee: &Expr<'tcx>) -> bool

Inspects the match scrutinee expression to determine whether the place it evaluates to may hold invalid data.

fn new_cx( &self, refutability: RefutableFlag, whole_match_span: Option<Span>, scrutinee: Option<&Expr<'tcx>>, scrut_span: Span, ) -> PatCtxt<'p, 'tcx>

fn analyze_patterns( &mut self, cx: &PatCtxt<'p, 'tcx>, arms: &[MatchArm<'p, 'tcx>], scrut_ty: Ty<'tcx>, ) -> Result<UsefulnessReport<'p, 'tcx>, ErrorGuaranteed>

fn check_let( &mut self, pat: &'p Pat<'tcx>, scrutinee: Option<ExprId>, span: Span, )

fn check_match( &mut self, scrut: ExprId, arms: &[ArmId], source: MatchSource, expr_span: Span, )

fn check_let_chain( &mut self, chain_refutabilities: Vec<Option<(Span, RefutableFlag)>>, whole_chain_span: Span, )

fn analyze_binding( &mut self, pat: &'p Pat<'tcx>, refutability: RefutableFlag, scrut: Option<&Expr<'tcx>>, ) -> Result<(PatCtxt<'p, 'tcx>, UsefulnessReport<'p, 'tcx>), ErrorGuaranteed>

fn is_let_irrefutable( &mut self, pat: &'p Pat<'tcx>, scrut: Option<&Expr<'tcx>>, ) -> Result<RefutableFlag, ErrorGuaranteed>

fn check_binding_is_irrefutable( &mut self, pat: &'p Pat<'tcx>, origin: &str, scrut: Option<&Expr<'tcx>>, sp: Option<Span>, )