Struct rustc_mir_build::thir::pattern::PatCtxt

struct PatCtxt<'a, 'tcx> {
    tcx: TyCtxt<'tcx>,
    param_env: ParamEnv<'tcx>,
    typeck_results: &'a TypeckResults<'tcx>,
}

Fields
tcx: TyCtxt<'tcx>param_env: ParamEnv<'tcx>typeck_results: &'a TypeckResults<'tcx>
Implementations

impl<'a, 'tcx> PatCtxt<'a, 'tcx>

pub(super) fn const_to_pat( &self, cv: ConstantKind<'tcx>, id: HirId, span: Span, check_body_for_struct_match_violation: Option<DefId> ) -> Box<Pat<'tcx>>

Converts an evaluated constant to a pattern (if possible). This means aggregate values (like structs and enums) are converted to a pattern that matches the value (as if you’d compared via structural equality).

impl<'a, 'tcx> PatCtxt<'a, 'tcx>

fn lower_pattern(&mut self, pat: &'tcx Pat<'tcx>) -> Box<Pat<'tcx>>

fn lower_range_expr( &mut self, expr: &'tcx Expr<'tcx> ) -> (PatKind<'tcx>, Option<Ascription<'tcx>>)

fn lower_pattern_range( &mut self, ty: Ty<'tcx>, lo: ConstantKind<'tcx>, hi: ConstantKind<'tcx>, end: RangeEnd, span: Span, lo_expr: Option<&Expr<'tcx>>, hi_expr: Option<&Expr<'tcx>> ) -> PatKind<'tcx>

fn normalize_range_pattern_ends( &self, ty: Ty<'tcx>, lo: Option<&PatKind<'tcx>>, hi: Option<&PatKind<'tcx>> ) -> Option<(ConstantKind<'tcx>, ConstantKind<'tcx>)>

fn lower_pattern_unadjusted(&mut self, pat: &'tcx Pat<'tcx>) -> Box<Pat<'tcx>>

fn lower_tuple_subpats( &mut self, pats: &'tcx [Pat<'tcx>], expected_len: usize, gap_pos: DotDotPos ) -> Vec<FieldPat<'tcx>>

fn lower_patterns(&mut self, pats: &'tcx [Pat<'tcx>]) -> Box<[Box<Pat<'tcx>>]>

fn lower_opt_pattern( &mut self, pat: &'tcx Option<&'tcx Pat<'tcx>> ) -> Option<Box<Pat<'tcx>>>

fn slice_or_array_pattern( &mut self, span: Span, ty: Ty<'tcx>, prefix: &'tcx [Pat<'tcx>], slice: &'tcx Option<&'tcx Pat<'tcx>>, suffix: &'tcx [Pat<'tcx>] ) -> PatKind<'tcx>

fn lower_variant_or_leaf( &mut self, res: Res, hir_id: HirId, span: Span, ty: Ty<'tcx>, subpatterns: Vec<FieldPat<'tcx>> ) -> PatKind<'tcx>

fn lower_path( &mut self, qpath: &QPath<'_>, id: HirId, span: Span ) -> Box<Pat<'tcx>>

Takes a HIR Path. If the path is a constant, evaluates it and feeds it to const_to_pat. Any other path (like enum variants without fields) is converted to the corresponding pattern via lower_variant_or_leaf.

fn lower_inline_const( &mut self, block: &'tcx ConstBlock, id: HirId, span: Span ) -> PatKind<'tcx>

Converts inline const patterns.

fn lower_lit(&mut self, expr: &'tcx Expr<'tcx>) -> PatKind<'tcx>

Converts literals, paths and negation of literals to patterns. The special case for negation exists to allow things like -128_i8 which would overflow if we tried to evaluate 128_i8 and then negate afterwards.