Module rustc_mir_build::build::matches

Code related to match expressions. These are sufficiently complex to warrant their own module and submodules. :) This main module includes the high-level algorithm, the submodules contain the details.

This also includes code for pattern bindings in let statements and function parameters.

Modules

simplify ğŸ”’

Simplifying Candidates

test ğŸ”’

util ğŸ”’

Structs

ArmHasGuard ğŸ”’

ArmHasGuard is a wrapper around a boolean flag. It indicates whether a match arm has a guard expression attached to it.

Ascription ğŸ”’

Indicates that the type of source must be a subtype of the user-given type user_ty; this is basically a no-op but can influence region inference.

Binding ğŸ”’

Candidate ğŸ”’

MatchPair ğŸ”’

Test ğŸ”’

A test to perform to determine which Candidate matches a value.

Enums

TestKind ğŸ”’

See Test for more.

Functions

traverse_candidate ğŸ”’

A depth-first traversal of the Candidate and all of its recursive subcandidates.