Trait rustc_mir_build::build::scope::DropTreeBuilder

trait DropTreeBuilder<'tcx> {
    // Required methods
    fn make_block(cfg: &mut CFG<'tcx>) -> BasicBlock;
    fn add_entry(cfg: &mut CFG<'tcx>, from: BasicBlock, to: BasicBlock);
}

A trait that determined how DropTree creates its blocks and links to any entry nodes.

Required Methods

fn make_block(cfg: &mut CFG<'tcx>) -> BasicBlock

Create a new block for the tree. This should call either cfg.start_new_block() or cfg.start_new_cleanup_block().

fn add_entry(cfg: &mut CFG<'tcx>, from: BasicBlock, to: BasicBlock)

Links a block outside the drop tree, from, to the block to inside the drop tree.