rustc_mir_build::thir::pattern::check_match
Function find_fallback_pattern_typo

fn find_fallback_pattern_typo<'tcx>(
    cx: &RustcPatCtxt<'_, 'tcx>,
    hir_id: HirId,
    pat: &Pat<'tcx>,
    lint: &mut UnreachablePattern<'_>,
)

Detect typos that were meant to be a const but were interpreted as a new pattern binding.